generate-*-api: fix on-disk tap migrations not being used

+++ Library/Homebrew/dev-cmd/generate-cask-api.rb
@@ -49,11 +49,11 @@ module Homebrew
       directories = ["_data/cask", "api/cask", "api/cask-source", "cask"].freeze
       FileUtils.rm_rf directories
       FileUtils.mkdir_p directories
-
-      File.write("api/cask_tap_migrations.json", JSON.dump(tap.tap_migrations))
     end
 
     Homebrew.with_no_api_env do
+      File.write("api/cask_tap_migrations.json", JSON.dump(tap.tap_migrations)) unless args.dry_run?
+
       Cask::Cask.generating_hash!
 
       tap.cask_files.each do |path|
